+++ riscv-macros/src/riscv_rt.rs
@@ -9,6 +9,7 @@ use syn::{
 /// Enum representing the supported runtime function attributes
 pub enum Fn {
     PostInit,
+    Entry,
 }
 
 impl Fn {
@@ -16,6 +17,12 @@ impl Fn {
     pub fn post_init(args: TokenStream, input: TokenStream) -> TokenStream {
         let errors = Self::PostInit.check_args_empty(args).err();
         Self::PostInit.quote_fn(input, errors)
+    }
+
+    /// Convenience method to generate the token stream for the `entry` attribute
+    pub fn entry(args: TokenStream, input: TokenStream) -> TokenStream {
+        let errors = Self::Entry.check_args_empty(args).err();
+        Self::Entry.quote_fn(input, errors)
     }
 
     /// Generate the token stream for the function with the given attribute
@@ -111,6 +118,10 @@ impl Fn {
         // Use this match to specify expected input arguments for different functions in the future
         match self {
             Self::PostInit => self.check_fn_args(inputs, &["usize"], errors),
+            #[cfg(not(feature = "u-boot"))]
+            Self::Entry => self.check_fn_args(inputs, &["usize", "usize", "usize"], errors),
+            #[cfg(feature = "u-boot")]
+            Self::Entry => self.check_fn_args(inputs, &["c_int", "*const *const c_char"], errors),
         }
     }
 
@@ -119,6 +130,7 @@ impl Fn {
         // Use this match to specify expected output types for different functions in the future
         match self {
             Self::PostInit => check_output_empty(output, errors),
+            Self::Entry => check_output_never(output, errors),
         }
     }
 
@@ -127,12 +139,17 @@ impl Fn {
         // Use this match to specify export names for different functions in the future
         let export_name = match self {
             Self::PostInit => Some("__post_init".to_string()),
+            Self::Entry => Some("main".to_string()),
         };
 
-        export_name.map(|name| {
-            quote! {
+        export_name.map(|name| match self {
+            Self::Entry => quote! {
+                // to avoid two main symbols when testing on host
+                #[cfg_attr(any(target_arch = "riscv32", target_arch = "riscv64"), export_name = #name)]
+            },
+            _ => quote! {
                 #[export_name = #name]
-            }
+            },
         })
     }
 
@@ -140,7 +157,7 @@ impl Fn {
     fn link_section(&self, _f: &ItemFn) -> Option<TokenStream2> {
         // Use this match to specify section names for different functions in the future
         let section_name: Option<String> = match self {
-            Self::PostInit => None,
+            Self::PostInit | Self::Entry => None,
         };
 
         section_name.map(|section| quote! {
@@ -247,5 +264,12 @@ fn check_output_empty(output: &ReturnType, errors: &mut Option<Error>) {
             }
             _ => combine_err(errors, Error::new(ty.span(), "return type must be ()")),
         },
+    }
+}
+
+/// Make sure the output type is `!` (never)
+fn check_output_never(output: &ReturnType, errors: &mut Option<Error>) {
+    if !matches!(output, ReturnType::Type(_, ty) if matches!(**ty, Type::Never(_))) {
+        combine_err(errors, Error::new(output.span(), "return type must be !"));
     }
 }